Code refactoring, styling
