+++ frappe/public/css/datatable.css
@@ -14,6 +14,8 @@
   top: 0;
   left: 0;
   background-color: white;
+  color: #8D99A6;
+  font-weight: bold;
 }
 .data-table-col {
   display: table-cell;
@@ -21,7 +23,16 @@
   white-space: nowrap;
 }
 .data-table-col > div {
-  padding: 5px;
+  padding: 10px;
+}
+.data-table-col.numeric {
+  text-align: right;
+}
+.data-table-col div {
+  border: 1px solid white;
+}
+.data-table-col.selected > div {
+  border: 1px solid #5E64FF;
 }
 .data-table-row {
   display: table-row;
