+++ frappe/public/less/datatable.less
@@ -2,7 +2,6 @@
 
 .data-table {
 	width: 100%;
-	// border: 1px solid @border-color;
 	position: relative;
 	overflow: auto;
 	font-size: @text-medium;
@@ -12,39 +11,40 @@
 .body-scrollable {
 	max-height: 500px;
 	overflow: auto;
-	// margin-top: 35px;
-
-
-	// th {
-	// 	white-space: nowrap;
-	// }
-
-	// td {
-	// 	max-height: 20px;
-	// }
 }
 
 .data-table-header {
 	position: absolute;
 	top: 0;
 	left: 0;
-	// width: 100%;
 	background-color: white;
-
-	// th {
-	// 	white-space: nowrap;
-	// }
+	color: @text-muted;
+	font-weight: bold;
 }
 
 .data-table-col {
 	display: table-cell;
 	border: 1px solid @border-color;
 	white-space: nowrap;
-	// padding: 5px;
 
 	& > div {
-		padding: 5px;
+		padding: 10px;
 	}
+
+	&.numeric {
+		text-align: right;
+	}
+
+	div {
+		border: 1px solid white;
+	}
+
+	&.selected > div {
+		border: 1px solid @brand-primary;
+	}
+	// &:focus {
+	// 	outline: 2px solid @brand-primary;
+	// }
 }
 
 .data-table-row {
